requirements.txt
langchain
langchain-cohere
cohere
ipykernel
python-dotenv
beautifulsoup4
html2text
clean-text
unidecode
chromadb
streamlit
